External link - refactor
Move external link utilities into their own lib file.
Remove obsolete function.
